# swagger.yaml
openapi: 3.0.3
info:
  title: Succulent Identifier API
  description: |
    REST API for the Succulent Identifier application. This API provides endpoints for:
    - Plant identification using machine learning
    - AI-powered chat assistance for plant care questions
    - Identification history management
    - Image serving

    Built with Go, PostgreSQL, OpenAI GPT-4o-mini, and PyTorch.
  version: 1.0.0
  contact:
    name: Succulent Identifier
    url: https://github.com/yourusername/succulent_identifier
servers:
  - url: http://localhost:8080
    description: Local development server (Backend API)
  - url: http://localhost:8000
    description: Local ML service
...
paths:
  /identify:
    post:
      tags:
        - Identification
      summary: Identify a succulent plant
      description: |
        Upload an image of a succulent plant to get species identification and care instructions.
        The API uses a confidence threshold (0.4) to determine whether to show species or genus-level results.
      operationId: identifyPlant
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - image
              properties:
                image:
                  type: string
                  format: binary
                  description: Image file (JPG or PNG, max 5MB)
      responses:
        '200':
          description: Successful identification
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentifyResponse'
              examples:
                high_confidence:
                  summary: High confidence (>= 0.4) - shows species
                  value:
                    id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                    plant:
                      genus: "Haworthia"
                      species: "Haworthia Zebrina"
                      confidence: 0.9468
                    care:
                      sunlight: "Bright, indirect light. Avoid direct sun."
                      watering: "Water when soil is completely dry (every 2-3 weeks)."
                      soil: "Well-draining cactus or succulent mix."
                      notes: "Hardy and easy to care for. Great for beginners."
                low_confidence:
                  summary: Low confidence (< 0.4) - shows genus only
                  value:
                    id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                    plant:
                      genus: "Haworthia"
                      species: ""
                      confidence: 0.35
                    care:
                      sunlight: "Bright, indirect light"
                      watering: "Water sparingly"
                      soil: "Well-draining mix"
                      notes: "Genus-level care information"
        '400':
          description: Bad request - invalid file or missing image
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Bad Request"
                message: "Invalid file type. Only JPG and PNG are supported."
        '500':
          description: Internal server error - ML service failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Internal Server Error"
                message: "Failed to communicate with ML service"
  /chat:
    post:
      tags:
        - Chat
      summary: Chat with AI about identified plant
      description: |
        Send a message to the AI chat assistant about an identified plant. The assistant has context
        about the plant's identification and care requirements.
      operationId: sendChatMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
            example:
              identification_id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
              message: "How often should I water this plant?"
      responses:
        '200':
          description: Successful chat response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
              example:
                message: "Based on the Haworthia zebrina identification, you should water when the soil is completely dry, typically every 2-3 weeks. These plants store water in their leaves, so they're drought-tolerant."
                message_id: "b2c3d4e5-f6a7-8901-bcde-f12345678901"
                timestamp: "2026-02-17T22:30:00Z"
        '400':
          description: Bad request - invalid identification_id or message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Identification not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error - OpenAI API failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /history:
    get:
      tags:
        - History
      summary: Get paginated list of identifications
      description: Retrieve a paginated list of past plant identifications
      operationId: getHistory
      parameters:
        - name: limit
          in: query
          description: Number of items to return (max 100)
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          description: Number of items to skip for pagination
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: Successful response with identification list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryListResponse'
              example:
                items:
                  - id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                    genus: "Haworthia"
                    species: "haworthia_zebrina"
                    confidence: 0.9468
                    image_path: "4cb08722-461a-4d6f-acd4-b06516cde3e8.jpg"
                    created_at: "2026-02-17T22:08:59Z"
                  - id: "b2c3d4e5-f6a7-8901-bcde-f12345678901"
                    genus: "Opuntia"
                    species: "opuntia_microdasys"
                    confidence: 0.9730
                    image_path: "5dc09833-572b-5e8f-bde5-g23456789012.jpg"
                    created_at: "2026-02-17T21:45:30Z"
                total: 4
                limit: 20
                offset: 0
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    PlantInfo:
      type: object
      properties:
        genus:
          type: string
          description: Plant genus name
          example: "Haworthia"
        species:
          type: string
          description: Plant species name (empty if confidence < threshold)
          example: "Haworthia Zebrina"
        confidence:
          type: number
          format: float
          description: Prediction confidence score (0-1)
          minimum: 0
          maximum: 1
          example: 0.9468
    CareInstructions:
      type: object
      properties:
        sunlight:
          type: string
          description: Sunlight requirements
          example: "Bright, indirect light. Avoid direct sun."
        watering:
          type: string
          description: Watering instructions
          example: "Water when soil is completely dry (every 2-3 weeks)."
        soil:
          type: string
          description: Soil recommendations
          example: "Well-draining cactus or succulent mix."
        notes:
          type: string
          description: Additional care notes
          example: "Hardy and easy to care for. Great for beginners."
    IdentifyResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identification ID
        plant:
          $ref: '#/components/schemas/PlantInfo'
        care:
          $ref: '#/components/schemas/CareInstructions'
    ChatRequest:
      type: object
      required:
        - identification_id
        - message
      properties:
        identification_id:
          type: string
          format: uuid
          description: ID of the plant identification
        message:
          type: string
          description: User's question or message
          example: "How often should I water this plant?"
    ChatResponse:
      type: object
      properties:
        message:
          type: string
          description: AI assistant's response
          example: "Water every 2-3 weeks when soil is completely dry."
        message_id:
          type: string
          format: uuid
          description: Unique message ID
        timestamp:
          type: string
          format: date-time
          description: Message timestamp
    HistoryItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        genus:
          type: string
        species:
          type: string
        confidence:
          type: number
          format: float
        image_path:
          type: string
          description: Image filename (not full path)
        created_at:
          type: string
          format: date-time
    HistoryListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/HistoryItem'
        total:
          type: integer
          description: Total number of identifications
        limit:
          type: integer
          description: Number of items per page
        offset:
          type: integer
          description: Current offset
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error type
          example: "Bad Request"
        message:
          type: string
          description: Error message
          example: "Invalid file type"